[#39] Rename next-session
