+++ nextjs-13/src/app/layout.tsx
@@ -1,10 +1,6 @@
-import { NextAuthProvider } from "./providers";
+import { NextAuthProvider } from "./session-providers";
 import App from "./app";
 
-export const metadata = {
-  title: 'Next.js',
-  description: 'Generated by Next.js',
-}
 
 export default function RootLayout({
   children,
